Disable fail-fast mode for CI (#712)
GitHub Actions by default cancels all other builds in a matrix when one fails,
meaning that we weren't getting a full report of which example repos suceeded
and which failed. By adding `fail-fast: false`, we'll now get a full report, and
retrying on an individual level if necessary should be easier.