Fix parse failure in TS when default-importing the name "type" (#722)
The token `from` is actually a `tt.name`, so we can use our existing lookahead
to distinguish that case. The line `import type from from './foo';` is invalid,
so we know that `from` after `type` is always a true `from` token, meaning that
`type` is just the imported name.

Also add a TODO comment to the babel test checker with a little more thoughts on
how to approach the remaining flagged issues.