Fix invalid output code when eliding TS star imports (#723)
From some manual testing, I noticed that the code:
```ts
import A, * as B from "./A";
console.log(A);
```
becomes this invalid code:
```ts
import A, from "./A";
console.log(A);
```
The `* as B` was being elided because it's treated as a type-only import, but in
this case we need to elide the comma as well since there's no concept of
"trailing commas" here. The code to do that is a little clunky, but I ended up
always removing the comma and using a boolean to track whether it should be
added back.