Update website UI to allow specifying all options (#736)
This PR makes it possible to specify almost all Sucrase options in the
playground/website UI. It should make it easier to develop the JSX automatic
transform and any future options, as well as test and report bugs that are
specific to options.

Currently, all options except `filePath` and `sourceMapOptions` are supported.
Those could potentially be added in the future, but seem less important for the
purposes of the website.

I was initially thinking about a modal for more advanced options, but I ended up
deciding instead on an expandable panel. This should make it easier to flip back
and forth between options to see how the output compares.

Since we need to include some custom code for each option anyway, I backtracked
a bit on the URL strategy from #735, and we now include the option name directly
as a URL arg. For now, the list of default values is also used as the list of
all options and their types, though this could be changed in the future if
necessary.